+++ Sources/DuckoCore/Services/ProfileService.swift
@@ -12,7 +12,7 @@ public final class ProfileService {
 
         public var errorDescription: String? {
             switch self {
-            case let .notConnected(id): notConnectedDescription(id)
+            case .notConnected: notConnectedDescription
             case let .invalidJID(string): "Invalid JID: \(string)"
             }
         }