+++ Sources/DuckoCore/Services/RosterService.swift
@@ -55,7 +55,7 @@ public final class RosterService {
 
         public var errorDescription: String? {
             switch self {
-            case let .notConnected(id): notConnectedDescription(id)
+            case .notConnected: notConnectedDescription
             case let .invalidJID(string): "Invalid JID: \(string)"
             }
         }